Create Handler service for join updates
We can separate responsibility by using the concept of Handlers to
move logic from the Controller moving forward.

To stay with the naming conventions of Join/Race/Countdown, we can
further see what's happening by re-naming this file.
